+++ src/interfaces.ts
@@ -79,7 +79,7 @@ export interface PostGraphileOptions<
   // may want to authenticate your users using sessions or similar. You can
   // pass some simple middlewares here that will be executed against the
   // websocket connection in order to perform authentication. We current only
-  // support express (not Koa) middlewares here.
+  // support Express (not Koa, Fastify, Restify, etc) middlewares here.
   /* @middlewareOnly */
   websocketMiddlewares?: Array<Middleware<Request, Response>>;
   // The default Postgres role to use. If no role was provided in a provided
@@ -238,6 +238,7 @@ export interface PostGraphileOptions<
   // the `Authorization` header, and signing JWT tokens you return in
   // procedures.
   jwtSecret?: jwt.Secret;
+  // The public key to verify the JWT when signed with RS265 or ES256 algorithms.
   jwtPublicKey?: jwt.Secret | jwt.GetPublicKeyOrSecret;
   // Options with which to perform JWT verification - see
   // https://github.com/auth0/node-jsonwebtoken#jwtverifytoken-secretorpublickey-options-callback