+++ src/postgraphile/http/createPostGraphileHttpRequestHandler.ts
@@ -17,7 +17,6 @@ import { pluginHookFromOptions } from '../pluginHook';
 import { HttpRequestHandler, mixed, CreateRequestHandlerOptions } from '../../interfaces';
 import setupServerSentEvents from './setupServerSentEvents';
 import withPostGraphileContext from '../withPostGraphileContext';
-import { Context as KoaContext } from 'koa';
 import LRU from '@graphile/lru';
 
 import chalk from 'chalk';
@@ -54,7 +53,8 @@ import favicon from '../../assets/favicon.ico';
 import baseGraphiqlHtml from '../../assets/graphiql.html';
 import { enhanceHttpServerWithWebSockets } from './subscriptions';
 import {
-  KoaNext,
+  CompatKoaContext,
+  CompatKoaNext,
   PostGraphileResponse,
   PostGraphileResponseKoa,
   PostGraphileResponseNode,
@@ -596,7 +596,20 @@ export default function createPostGraphileHttpRequestHandler(
     'eventStreamRouteHandler',
     async function eventStreamRouteHandler(res: PostGraphileResponse) {
       try {
-        const req = res.getNodeServerRequest();
+        // You can use this hook either to modify the incoming request or to tell
+        // PostGraphile not to handle the request further (return null). NOTE: if
+        // you return `null` from this hook then you are also responsible for
+        // terminating the request however your framework handles that (e.g.
+        // `res.send(...)` or `next()`).
+        const req = pluginHook(
+          'postgraphile:http:eventStreamRouteHandler',
+          res.getNodeServerRequest(),
+          { options, response: res },
+        );
+        if (req == null) {
+          return;
+        }
+
         // Add our CORS headers to be good web citizens (there are perf
         // implications though so be careful!)
         //
@@ -622,7 +635,19 @@ export default function createPostGraphileHttpRequestHandler(
   const faviconRouteHandler = neverReject('faviconRouteHandler', async function faviconRouteHandler(
     res: PostGraphileResponse,
   ) {
-    const req = res.getNodeServerRequest();
+    // You can use this hook either to modify the incoming request or to tell
+    // PostGraphile not to handle the request further (return null). NOTE: if
+    // you return `null` from this hook then you are also responsible for
+    // terminating the request however your framework handles that (e.g.
+    // `res.send(...)` or `next()`).
+    const req = pluginHook('postgraphile:http:faviconRouteHandler', res.getNodeServerRequest(), {
+      options,
+      response: res,
+    });
+    if (req == null) {
+      return;
+    }
+
     // If this is the wrong method, we should let the client know.
     if (!(req.method === 'GET' || req.method === 'HEAD')) {
       res.statusCode = req.method === 'OPTIONS' ? 200 : 405;
@@ -648,7 +673,19 @@ export default function createPostGraphileHttpRequestHandler(
   const graphiqlRouteHandler = neverReject(
     'graphiqlRouteHandler',
     async function graphiqlRouteHandler(res: PostGraphileResponse) {
-      const req = res.getNodeServerRequest();
+      // You can use this hook either to modify the incoming request or to tell
+      // PostGraphile not to handle the request further (return null). NOTE: if
+      // you return `null` from this hook then you are also responsible for
+      // terminating the request however your framework handles that (e.g.
+      // `res.send(...)` or `next()`).
+      const req = pluginHook('postgraphile:http:graphiqlRouteHandler', res.getNodeServerRequest(), {
+        options,
+        response: res,
+      });
+      if (req == null) {
+        return;
+      }
+
       if (firstRequestHandler) firstRequestHandler(req);
 
       // If using the incorrect method, let the user know.
@@ -687,7 +724,19 @@ export default function createPostGraphileHttpRequestHandler(
   const graphqlRouteHandler = neverReject('graphqlRouteHandler', async function graphqlRouteHandler(
     res: PostGraphileResponse,
   ) {
-    const req = res.getNodeServerRequest();
+    // You can use this hook either to modify the incoming request or to tell
+    // PostGraphile not to handle the request further (return null). NOTE: if
+    // you return `null` from this hook then you are also responsible for
+    // terminating the request however your framework handles that (e.g.
+    // `res.send(...)` or `next()`).
+    const req = pluginHook('postgraphile:http:graphqlRouteHandler', res.getNodeServerRequest(), {
+      options,
+      response: res,
+    });
+    if (req == null) {
+      return;
+    }
+
     if (firstRequestHandler) firstRequestHandler(req);
 
     // Add our CORS headers to be good web citizens (there are perf
@@ -1033,8 +1082,8 @@ export default function createPostGraphileHttpRequestHandler(
     // `koa` middleware.
     if (isKoaApp(a, b)) {
       // Set the correct `koa` variable names…
-      const ctx = a as KoaContext;
-      const next = b as KoaNext;
+      const ctx = a as CompatKoaContext;
+      const next = b as CompatKoaNext;
       const responseHandler = new PostGraphileResponseKoa(ctx, next);
 
       // Execute our request handler. If an error is thrown, we don’t call